Moved DTO classes into model/dto
