+++ backend/src/main/java/com/vybe/backend/model/dto/AdminDTO.java
@@ -1,4 +1,4 @@
-package com.vybe.backend.DTO;
+package com.vybe.backend.model.dto;
 
 import com.vybe.backend.model.entity.Admin;
 import lombok.AllArgsConstructor;
@@ -10,11 +10,10 @@ import lombok.Setter;
 @NoArgsConstructor
 @Getter
 @Setter
-public class AdminCreationDTO {
+public class AdminDTO {
     private String username;
-    private String password;
 
-    public Admin toAdmin() {
-        return new Admin(username, password);
+    public AdminDTO(Admin admin) {
+        this.username = admin.getUsername();
     }
 }
