+++ backend/src/main/java/com/vybe/backend/model/dto/VenueDTO.java
@@ -1,4 +1,4 @@
-package com.vybe.backend.DTO;
+package com.vybe.backend.model.dto;
 
 import com.vybe.backend.model.entity.*;
 import lombok.AllArgsConstructor;
